Move select reservation to separate controller
